feat(mixins): move createTheme export for no circular dep with Styles
