+++ Styles/Materials.js
@@ -17,9 +17,38 @@
  */
 
 import lng from '@lightningjs/core';
-import { CORNER_RADIUS } from './Styles';
 
 export const spacing = multiplier => 8 * multiplier;
+
+export const ALPHA = {
+  overlays: {
+    background: 72,
+    bubble: 56,
+    shelf: 56
+  },
+  text: {
+    primary: 100,
+    secondary: 72,
+    tertiary: 56,
+    inactive: 40
+  }
+};
+
+export const CORNER_RADIUS = {
+  large: 24,
+  medium: 16,
+  small: 8,
+  xsmall: 4
+};
+
+/**
+ * Amount to blur shaders on components like the Shelf.
+ *
+ * The cap for Fast Blur component is 4, but Sketch files use "50"
+ * The higher number is also necessary for when requesting a blurred version
+ * of an image from a service using something like image magick.
+ */
+export const BLUR_AMOUNT = 4;
 
 function glow({
   w,
